Organize the code in packages
