+++ spring-5-data-reactive/src/main/java/com/baeldung/r2dbc/R2dbcApplication.java
@@ -1,22 +1,11 @@
 package com.baeldung.r2dbc;
 
-import io.r2dbc.h2.H2ConnectionConfiguration;
-import io.r2dbc.h2.H2ConnectionFactory;
-import lombok.AllArgsConstructor;
-import lombok.Data;
-import lombok.NoArgsConstructor;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.data.annotation.Id;
-import org.springframework.data.r2dbc.config.AbstractR2dbcConfiguration;
-import org.springframework.data.r2dbc.repository.config.EnableR2dbcRepositories;
-import org.springframework.data.r2dbc.repository.query.Query;
-import org.springframework.data.repository.reactive.ReactiveCrudRepository;
-import reactor.core.publisher.Flux;
+import org.springframework.context.annotation.ComponentScan;
 
 @SpringBootApplication
+@ComponentScan(basePackages = "com.baeldung.r2dbc")
 public class R2dbcApplication {
 
     public static void main(String[] args) {
@@ -24,37 +13,3 @@ public class R2dbcApplication {
     }
 
 }
-
-interface PlayerRepository extends ReactiveCrudRepository<Player, Integer> {
-
-    @Query("select id, name, age from player where name = $1")
-    Flux<Player> findAllByName(String name);
-
-    @Query("select * from player where age = $1")
-    Flux<Player> findByAge(int age);
-}
-
-
-@Data
-@NoArgsConstructor
-@AllArgsConstructor
-class Player {
-    @Id
-    Integer id;
-    String name;
-    Integer age;
-}
-
-@Configuration
-@EnableR2dbcRepositories("org.springframework.data.r2dbc.repository.support.R2dbcRepositoryFactory.class")
-class R2DBCConfiguration extends AbstractR2dbcConfiguration {
-    @Bean
-    public H2ConnectionFactory connectionFactory() {
-        return new H2ConnectionFactory(
-                H2ConnectionConfiguration.builder()
-                        .url("mem:testdb;DB_CLOSE_DELAY=-1;TRACE_LEVEL_FILE=4")
-                        .username("sa")
-                        .build());
-    }
-}
-
