+++ spring-5-data-reactive/src/main/java/com/baeldung/r2dbc/model/Player.java
@@ -1,0 +1,16 @@
+package com.baeldung.r2dbc.model;
+
+import lombok.AllArgsConstructor;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+import org.springframework.data.annotation.Id;
+
+@Data
+@NoArgsConstructor
+@AllArgsConstructor
+public class Player {
+    @Id
+    Integer id;
+    String name;
+    Integer age;
+}
