+++ spring-5-data-reactive/src/test/java/com/baeldung/r2dbc/R2dbcApplicationUnitTest.java
@@ -1,6 +1,8 @@
 package com.baeldung.r2dbc;
 
 
+import com.baeldung.r2dbc.model.Player;
+import com.baeldung.r2dbc.repository.PlayerRepository;
 import io.r2dbc.h2.H2ConnectionFactory;
 import org.junit.Before;
 import org.junit.Test;
